34019:修复当输入小数时，自动转整型的问题

+++ WebContent/mobile-v3/js/deposit/BaseDeposit.js
@@ -22,7 +22,7 @@ var BaseDeposit = function () {
         if(val.trim().length<1){
             document.getElementById("result.rechargeAmount").value = options.mone;
         }else {
-            document.getElementById("result.rechargeAmount").value = parseInt(val) + options.mone;
+            document.getElementById("result.rechargeAmount").value = parseFloat(val) + options.mone;
         }
     };
